// src/types.ts
export type IntegrationSource = 'Facebook' | 'Google Ads' | 'Website' | 'Cloud Telephony' | 'Email Marketing' | 'SMS Marketing';
export interface FieldMapping {
    sourceField: string;
    crmField: string;
}

export interface ConnectedFacebookPage {
    id: string;
    name: string;
    subscribed: boolean;
    forms: { id: string; name: string; fields: string[] }[];
}

export interface ConnectedFacebookAccount {
    id: string;
    name: string;
    accessToken: string;
    pages: ConnectedFacebookPage[];
}
// ...
export interface ConnectedGoogleAccount {
    id: string;
    name: string;
}

export interface ConnectedWebsite {
    id: string;
    url: string;
}

export interface IntegrationSettings {
    source: IntegrationSource;
    isConnected: boolean;
    fieldMappings: FieldMapping[];
    // Add specific connection details
    connectedAccounts?: (ConnectedGoogleAccount | ConnectedFacebookAccount)[]; // Generic container
    connectedWebsites?: ConnectedWebsite[];
    organizationId: string;
}